refactor: renamed `utils` package to `util`
